Font rasterization processing
This is a C program that plots a charset in a given vector font onto a
Cairo surface and exports that surface then as a JSON file. Then a
Python script picks up that JSON file and creates C code from it (right
now it only generates PNM so you can see how good your rasterization
looks). It could all have been Python, but the Python Cairo bindings do
not support the surface.get_data() call (yet?). So unfortunately this is
the result.