Script to decided between vector or raster images
In the SVG, both links to the PNG files are present as well as the
vector drawings underneath. This hacks Python script pries the SVG apart
and creates two resulting SVGs, one which sets opacity for all vector
components to zero and the other which sets opacity for all raster
components to zero, effectively giving a image-only and a no-images SVG
that can be viewed.