Further instrument support
NAV VHF ADF and DME support. Bit field mask for lights misbehaves in the
same manner that BCD32-encoded ADF frequency did. Something seems to be
wrong in the interpretation of the integer bitfields. Switching to
floats fixed it for ADF, but I want to access all lights in one bitfield
(not access the single 64 bit "bool" values one-by-one). Hmmmm....