Conversion of FSConnection to C++
Changed the FSConnection to C++, since it really makes sense in this
part of the code. Also added an "Emulator" class (that allows testing
without a flight sim) and made the code compile both under Linux and
Windows (with the hope that I can make the Linux version also compile
for X-Plane 11).