Bounded buffer implementation of RS232
This doesn't work quite yet. But the idea is that RS232 data is cached
in a bounded buffer and periodically sent out IRQ-driven instead of
waiting for finishing up on every printf() call. This should allow use
of debug messages in places where it wouldn't otherwise be possible for
timing reasons.