Improved (hard) fault handler
Improved fault handler implementation. The stub is implemented in pure
assembly and manages to consistently pass the exact state at the time of
the IRQ to the fault handler. The handler then can dump all information
of interest (register set, especially the PC). It also dumps additional
bus fault register contents and decodes bitfields. This is useful
because for memory faults (e.g., access to invalid memory) the PC may
already have advanced one or several instructions when the fault occurs.
I.e., while it is a synchronous trap, it still shows asynchronous
behavior (this is called an "imprecise" fault in ARM terminology).
Therefore it's very useful to know what kind of fault you're dealing
with (so that debugging does not happen in a place that may be
innocent).