Fix deadlock issue for RS232 and new Python GDB tool
The deadlock issue that could potentially have occured when printf() was
called from within an ISR, leading to a TX buffer overflow, is now
solved. If the transmitchar() routine now sees that putting the send
character into the buffer failed (because it's full), it now raises the
IRQ priority of the USART2. That way, the USART2 ISR can preemt the
currently running ISR, making space inside the buffer. In normal
operation (i.e., when the buffer is not full) the ISR priority of the
USART2 is always lowest (i.e., it will usually *not* preemt other ISRs,
only when necessary).

Debugging of this issue took so long because mysteriously the ISR
priority of SysTick kept changing. This is because the cube code
annoyingly fiddles with it in the LL_USB_Init() routine. Therefore we
adapt bootstrp.sh to remove this absurd priority changing of
SysTick.Prio = 0.

In the future, we should remove all hardware-initializing code from
USB_LL_Init() and move it all to init.c. Which should also be cleaned up
to use the symbols from pinmap.h instead of raw pin numbers.

To debug this issue, I wrote some Python integration for gdb which
allows dumping of all System Handler states and priorities (inside the
System Control Block of the Cortex-M) and also all IRQ states of the
user IRQs (handled by the NVIC). Pretty useful, can be invoked by
calling "cm" currently. In the future, I might add more commands to this
facility (and rename the current one to respond to "cm irq" or something
like that).