Make stackpainting depend on link-time defines
Previously, the linker script defined _sram and _eram, but these were
not used within the stack painting. Instead, duplicate code in header
files defined the origin of RAM and the length of RAM again. This is
obviously crap -- now we rely on the authoritative values that are found
in the linker script.