Create script to build Travis-CI package
Travis-CI needs the gcc pre-built including with some headers that
you'll have to accept the license of. Therefore, these files are not
included in this git repo. Provided that you've agreed to all the
licensese involved, this script allows to create a Travis CI bundle.
I've also used it to create a new bundle that includes gcc-7.2.0.