Smarter, accurate temperature sensor calculation
The stability of the reference voltage has a great influence on the
outcome of the temperature sensor, much greater than the actual
calibration values for the temperature curves compared to the default
values from the datasheet. Unfortunately, the STM32DISCOVERY has a
series diode between the 3.3V LD3985 LDO, namely a BAT60J. Forward drop
across this ranges from 0.28V to 0.58V -- in other words, we don't know
Vdd (and Vref) at the STM32F407.

Therefore, we use the following approach: We enable Vrefint and measure
it. From it we assume that the original bandgap reference was 1.21 V and
can therefore infer Vref (which, for the STM32DISCOVERY, is identical to
Vdd).

We now use the calibration values at 30°C and 110°C that are stored in
ROM. The recorded values however are recorded at 3.3V. Therefore, we
calculate the 30°C and 110°C *voltages* (assuming a stable 3.3V
reference at factory calibration). With the previously determined Vref,
we can calculate Vtemp from ADU_temp accurately, then use linear
interpolation with the two calibration basepoints to calculate the
temperature.

The overall formula for this was calculated using sage and some
rearrangements have been made to it to lessen the impact of integer
division rounding. It gives good accuracy over the wide temperature
range and never needs to use 64-bit arithmetic for it (i.e., it's
*fast*).