More robust RS232 output logic
Previously, we waited for the TXC IRQ, but actually TXE (transmit data
register empty) is sufficient for us. Also replace the somewhat broken
mechanism of having a software flag "tx_in_progress" by inspecting the
actual USART register contents. This allows us to repeatedly call
send_data() and have it trigger an transmission if needed.

This solves the problem that large outputs of data from the USART2 RXC
ISR would deadlock the MCU. However, it introduced the new problem that
apparently, data is not transmitted reliably on bootup. Debug messages
only appear after a few returns are sent to the MCU.