Some more DMA debugging and IOMux pin mapping
Mapped out all IOMux pins, all are working and wired according to the
pinmap in the iomux.txt file.

Now when something happens in debug mode on the IOMux inputs, we can
resolve the name of the corresponding input. This is kind of slow and
lead to the discovery of an issue: DMA transfers to the IOMux SPI easily
"starve" when the system is under load (e.g., when putting it in
iomux-in debug mode and turning the rotary encoders quickly). Lots of
messages are generated (and memory for the RS232 buffer is written),
leading to the DMA of SPI3 to crash: While DMA1_Stream5 exits gracefully
(i.e., the TC ISR is executed and the DMA1_Stream5 is disabled),
DMA1_Stream2 (the RX stream) simply "hangs" while still enabled. Not
sure at the moment what the root cause for this behavior is. No DMA
error flags are set for either Stream2 or Stream5 and the SPI3 flags
also look fine (for all of the above, I've added debugging code).