Fixed hanging DMA issue
After a tip from @frankbuss, it appears that the hanging DMA issue
stemmed from the fact that the DMA_TC IRQ was pending too long
(especially because it couldn't preempt the debugging IRQ of RS232).
Instead of the ugly long-running ISRs, now implement a smallish ISR/DSR
model and only do the absolutely necessary tasks inside the ISR (such as
clearing the DMA request), and all the rest inside the DSR (such as
handling the received data). With this change, ISR preemtion should not
be necessary anymore, but still do it for good measure: Let the DMA IRQs
preempt all other IRQs if necessary.