Fix DMA issue for slow SPI transfers
When the clock rate of the SPI was decreased (e.g., to a 1/32 prescaler
instead of 1/8), there were display issues (flickering and/or wandering
of the image). The root cause of this was traced to be that the ISR
that's responsible for the DMA TC (transfer complete) is actually the
SPI TX (MOSI) ISR. We never need to read data from the display, but the
fix for this issue is to nevertheless introduce a MISO DMA request that
also fires a TC IRQ. The point is that the SPI TX DMA ISR may not
deassert CS because while the data has been transferred to the SPI
peripheral, the byte hasn't been clocked out yet (leading in the loss of
the last byte).  However, if we introduce a dummy read DMA transfer,
that TC will complete after the last byte has been fully clocked out/in
and we deassert CS in that ISR. For fast clock speeds, this bug didn't
show up because the transfer was so fast that it completed in the
meantime (i.e., when the ISR prologue was already running).