Refactoring of instrument code
Replace many booleans by bitfield for convenience; this also makes
transferring data to the host easier, since we can just copy it as-is.
Also refactor the whole data structure of the instrument_state_t
structure and divide it into parts that are internal (such as timeouts
or currently active screens) and external (state that is actively
transmitted to the host PC).