Re-implement setting of values by host
The new refactoring enables us to implement setting values by the host
(i.e., the connected PC) efficiently because many of the data structures
that the flight panel uses internally are used for the actual USB
transfer identically. This allows efficient setting and comparing of
current data vs. received data and making the right choices depending on
that (e.g., redrawing certain displays, but not others).