Put USB alternate function mapping and pin init in our code
Previously, this was re-inited by the CubeF4 code, but now we do the
assignment of USB AF and activation of the USB unit into our own code.
Currently, the relevant portion in CubeF4 is disabled manually, but we
need to disable this as well inside the sed commands of bootstrap.sh.