Change 30pt font to Inconsolata
VCR OSD Mono looks wonderful at 20pt, but horrible at 30pt. Inconsolata
looks a lot better in that regard. Change 30pt font to Inconsolata and
add several improvements on the font generation process. In particular,
allow command-line patching and custom properties of glyphs and
consolidate the "morse" font to become a "symbol" font (because we need
the plusminus sign for the vertical speed indication).